Remove cacheKey from cache promise, move hfp grouping function to a separate helper.
